Create Working CNN for images
Hey guys, I used Mike's original code and created a working CNN with all of the images. My Macbook pro cannot actually run this to completion (ETA of 3 hours) but I wanted to post this here if anyone else wanted to work with it. You'll have to truncate the amount of images you use and I haven't looked at how to do that yet. You also have to create separate folder locations for this to work, I'm not entirely sure why. We can zoom and I can explain more if you cant get it to work/have questions. This should be a base for us to be able to add new CV features or whatever else going forward