CNN Messidor
This is an updated CNN using messidor 2 dataset with the pictures and labels I uploaded to the drive. You'll have to change the directory, but this is my attempt of using AlexNet plus mixing some activations and adding label weights. Label weights might still need some work though becuase it is not currently set to batch, just weights the entire dataset.